requirements.txt
aioredis==1.2.0
amqp==2.4.2
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.2.1
Automat==0.7.0
bcrypt==3.1.6
billiard==3.5.0.5
celery==4.2.1
certifi==2019.3.9
cffi==1.12.2
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.6.1
daphne==2.2.5
DingtalkChatbot==1.3.0
Django==2.1.7
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-redis==4.10.0
django-timezone-field==3.0
djangorestframework==3.9.2
djangorestframework-stubs==0.3.0
et-xmlfile==1.0.1
gevent==1.4.0
greenlet==0.4.15
hiredis==1.0.0
humanfriendly==4.18
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
jdcal==1.4
kombu==4.4.0
msgpack==0.6.1
mysql-replication==0.18
mysqlclient==1.3.13
openpyxl==2.6.1
paramiko==2.4.2
Pillow==5.4.1
pyasn1==0.4.5
pycparser==2.19
PyHamcrest==1.9.0
PyMySQL==0.9.3
PyNaCl==1.3.0
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2018.9
redis==3.2.0
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.1
vine==1.2.0
zope.interface==4.6.0
